backend/requirements.txt
# Django & DRF
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5
drf-spectacular>=0.27,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment
python-decouple>=3.8,<4.0

# File handling
Pillow>=10.0,<11.0

# AI / ML
scikit-learn>=1.3,<2.0
langchain>=0.1,<1.0
langchain-community>=0.0.10
sentence-transformers>=2.2,<3.0
faiss-cpu>=1.7,<2.0

# PDF / Doc parsing
PyPDF2>=3.0,<4.0
python-docx>=1.0,<2.0

# NLP
nltk>=3.8,<4.0

# OCR
pytesseract>=0.3,<1.0

# Async Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0

# AI Concept Simplifier (OpenAI)
openai>=1.0,<3.0

# Utilities
python-magic>=0.4.27
gunicorn>=21.2,<22.0
whitenoise>=6.5,<7.0
